renaming Home to Starred for better representation
