+++ src/containers/DownloadContainer/DownloadContainer.js
@@ -12,7 +12,7 @@ const omitDeep = require('omit-deep-lodash')
 const videoSuffix = '.mp4' // Currently, classtrascribe only uses mp4 as their video format
 
 /**
- * Contains the home screen of the application. Lists starred courses and gives the user the ability
+ * Contains the starred screen of the application. Lists starred courses and gives the user the ability
  * to search for courses. Clicking on a course shows the playlists for it.
  * @param {Object} navigation A stack navigator that contains a video screen
  */
